Refactor radfield estimators to solve circular import
