+++ tardis/montecarlo/estimators/radfield_mc_estimators.py
@@ -1,18 +1,7 @@
-from math import exp
-
 import numpy as np
-from numba import float64, int64, njit
+from numba import float64, int64
 from numba.experimental import jitclass
-
-from tardis.montecarlo import montecarlo_configuration as nc
-from tardis.montecarlo.montecarlo_numba import (
-    njit_dict_no_parallel,
-)
-from tardis.montecarlo.montecarlo_numba.numba_config import KB, H
-from tardis.transport.frame_transformations import (
-    calc_packet_energy,
-    calc_packet_energy_full_relativity,
-)
+from numba.typed import List
 
 
 def initialize_estimator_statistics(tau_sobolev_shape, gamma_shape):
@@ -131,109 +120,21 @@ class RadiationFieldMCEstimators:
             other.photo_ion_estimator_statistics
         )
 
+    def create_list(self, number):
+        estimator_list = List()
 
-@njit(**njit_dict_no_parallel)
-def update_base_estimators(
-    r_packet, distance, estimator_state, comov_nu, comov_energy
-):
-    """
-    Updating the estimators
-    """
-    estimator_state.j_estimator[r_packet.current_shell_id] += (
-        comov_energy * distance
-    )
-    estimator_state.nu_bar_estimator[r_packet.current_shell_id] += (
-        comov_energy * distance * comov_nu
-    )
-
-
-@njit(**njit_dict_no_parallel)
-def update_bound_free_estimators(
-    comov_nu,
-    comov_energy,
-    shell_id,
-    distance,
-    estimator_state,
-    t_electron,
-    x_sect_bfs,
-    current_continua,
-    bf_threshold_list_nu,
-):
-    """
-    Update the estimators for bound-free processes.
-
-    Parameters
-    ----------
-    comov_nu : float
-    comov_energy : float
-    shell_id : int
-    distance : float
-    numba_estimator : tardis.montecarlo.montecarlo_numba.numba_interface.Estimators
-    t_electron : float
-        Electron temperature in the current cell.
-    x_sect_bfs : numpy.ndarray, dtype float
-        Photoionization cross-sections of all bound-free continua for
-        which absorption is possible for frequency `comov_nu`.
-    current_continua : numpy.ndarray, dtype int
-        Continuum ids for which absorption is possible for frequency `comov_nu`.
-    bf_threshold_list_nu : numpy.ndarray, dtype float
-        Threshold frequencies for photoionization sorted by decreasing frequency.
-    """
-    # TODO: Add full relativity mode
-    boltzmann_factor = exp(-(H * comov_nu) / (KB * t_electron))
-    for i, current_continuum in enumerate(current_continua):
-        photo_ion_rate_estimator_increment = (
-            comov_energy * distance * x_sect_bfs[i] / comov_nu
-        )
-        estimator_state.photo_ion_estimator[
-            current_continuum, shell_id
-        ] += photo_ion_rate_estimator_increment
-        estimator_state.stim_recomb_estimator[current_continuum, shell_id] += (
-            photo_ion_rate_estimator_increment * boltzmann_factor
-        )
-        estimator_state.photo_ion_estimator_statistics[
-            current_continuum, shell_id
-        ] += 1
-
-        nu_th = bf_threshold_list_nu[current_continuum]
-        bf_heating_estimator_increment = (
-            comov_energy * distance * x_sect_bfs[i] * (1 - nu_th / comov_nu)
-        )
-        estimator_state.bf_heating_estimator[
-            current_continuum, shell_id
-        ] += bf_heating_estimator_increment
-        estimator_state.stim_recomb_cooling_estimator[
-            current_continuum, shell_id
-        ] += (bf_heating_estimator_increment * boltzmann_factor)
-
-
-@njit(**njit_dict_no_parallel)
-def update_line_estimators(
-    radfield_mc_estimators,
-    r_packet,
-    cur_line_id,
-    distance_trace,
-    time_explosion,
-):
-    """
-    Function to update the line estimators
-
-    Parameters
-    ----------
-    estimators : tardis.montecarlo.montecarlo_numba.numba_interface.Estimators
-    r_packet : tardis.montecarlo.montecarlo_numba.r_packet.RPacket
-    cur_line_id : int
-    distance_trace : float
-    time_explosion : float
-    """
-    if not nc.ENABLE_FULL_RELATIVITY:
-        energy = calc_packet_energy(r_packet, distance_trace, time_explosion)
-    else:
-        energy = calc_packet_energy_full_relativity(r_packet)
-
-    radfield_mc_estimators.j_blue_estimator[
-        cur_line_id, r_packet.current_shell_id
-    ] += (energy / r_packet.nu)
-    radfield_mc_estimators.Edotlu_estimator[
-        cur_line_id, r_packet.current_shell_id
-    ] += energy
+        for i in range(number):
+            estimator_list.append(
+                RadiationFieldMCEstimators(
+                    np.copy(self.j_estimator),
+                    np.copy(self.nu_bar_estimator),
+                    np.copy(self.j_blue_estimator),
+                    np.copy(self.Edotlu_estimator),
+                    np.copy(self.photo_ion_estimator),
+                    np.copy(self.stim_recomb_estimator),
+                    np.copy(self.bf_heating_estimator),
+                    np.copy(self.stim_recomb_cooling_estimator),
+                    np.copy(self.photo_ion_estimator_statistics),
+                )
+            )
+        return estimator_list
